# src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: SpongeblogSP API
  description: Spongeblog blogging API
  version: "1.0.0"
# array of all schemes that your API supportsf
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
parameters:
...
  TagEdit:
    name: tag-edit
    in: body
    schema:
      $ref: '#/definitions/TagDTO'
    required: true
...
  IdTag:
    name: tag-id
    in: path
    description: Unique identifier of a Tag;
    required: true
    type: integer
    format: int64
...
paths:
...
  /api/v1/tags:
    get:
      summary: All Tags
      description: Returns all Tags.
      tags:
        - Tags
      responses:
        200:
          description: Tags.
          schema:
            type: array
            items:
              $ref: '#/definitions/TagDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    post:
      summary: Create Tag
      description: Creates new Tag entry.
      parameters:
        - $ref: '#/parameters/TagEdit'
      tags: 
        - Tags
      responses:
        201:
          description: Created Tag.
          schema:
            $ref: '#/definitions/TagDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    put:
      summary: Update Tag
      description: Updated a Tag entry.
      parameters:
        - $ref: '#/parameters/TagEdit'
      tags: 
        - Tags
      responses:
        200:
          description: Tag updated.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/tags/{tag-id}:
    get:
      summary: Get Tag
      description: Fetches Tag with defined Id.
      parameters:
        - $ref: '#/parameters/IdTag'
      tags: 
        - Tags
      responses:
        200:
          description: Tag with id.
          schema:
            $ref: '#/definitions/TagPostsDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    delete:
      summary: Delete Tag
      description: Deletes a Tag with defined Id.
      parameters:
        - $ref: '#/parameters/IdTag'
      tags: 
        - Tags
      responses:
        200:
          description: Tag deleted.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/tags/rewind/{tag-id}:
    put:
      summary: Rewind Tag
      description: Restores previous Tag entry.
      parameters:
        - $ref: '#/parameters/IdTag'
      tags: 
        - Tags
      responses:
        200:
          description: Previous Tag restored.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
...
definitions:
...
  TagDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier.
      name:
        type: string
        description: Tag name.
  TagPostsDTO:
    type: object
    properties:
      name:
        type: string
        description: Tag name.
      posts:
        type: string
        description: Link to related Posts.
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string